TensorTrainApprox

Documentation for TensorTrainApprox.

TensorTrainApprox.DMRGcross

TensorTrainApprox.DMRGcross

TensorTrainApprox.DMRGcross_threaded

TensorTrainApprox.TTderivative

TensorTrainApprox.TTevaluate

TensorTrainApprox.TTevaluate

TensorTrainApprox.TTevaluate

TensorTrainApprox.TTevaluate

TensorTrainApprox.TTgradient

TensorTrainApprox.TTintegrate_GC

TensorTrainApprox.TTintegrate_GC

TensorTrainApprox.TTintegrate_GH

TensorTrainApprox.TTintegrate_GL

TensorTrainApprox.TTinterp

TensorTrainApprox.TTsvd

TensorTrainApprox.constantTT

TensorTrainApprox.constantTT

TensorTrainApprox.createCTT

TensorTrainApprox.createCTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createPTT

TensorTrainApprox.decompress

TensorTrainApprox.decompress_threaded

TensorTrainApprox.frobenius

TensorTrainApprox.frobenius

TensorTrainApprox.ind2sub

TensorTrainApprox.ind2sub

TensorTrainApprox.randomTT

TensorTrainApprox.randomTT

TensorTrainApprox.times_cores

TensorTrainApprox.tsvd

TensorTrainApprox.DMRGcross — Method

Compute a tensor train approximation of a dense d-dimensional array. DMRG tensor compression based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross — Method

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_threaded — Method

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.TTderivative — Method

Compute the derivative of a Chebyshev tensor train with respect to a variable.

TensorTrainApprox.TTevaluate — Method

Evaluates a tensor train at a vector of index points of the dense array.

TensorTrainApprox.TTevaluate — Method

Evaluates a tensor train at an index point of the dense array.

TensorTrainApprox.TTevaluate — Method

Evaluate a functional tensor train at a vector of points.

TensorTrainApprox.TTevaluate — Method

Evaluate a functional tensor train at a point.

TensorTrainApprox.TTgradient — Method

Compute the gradient vector of a Chebyshev tensor train.

TensorTrainApprox.TTintegrate_GC — Method

Intrgrate a tensor train using Gauss-Chebyshev quadrature.

TensorTrainApprox.TTintegrate_GC — Method

Integrate a functional tensor train using Gauss-Chebyshev quadrature.

TensorTrainApprox.TTintegrate_GH — Method

Integrate a tensor train using Gauss-Hermite quadrature.

TensorTrainApprox.TTintegrate_GL — Method

Integrate a tensor train using Gauss-Legendre quadrature.

TensorTrainApprox.TTinterp — Method

Create an interpolating function from a functional tensor train.

TensorTrainApprox.TTsvd — Method

Compute a tensor train approximation of a dense d-dimensional array. Tensor compression based on Oseledets and Tyrtyshnikov (2010).

TensorTrainApprox.constantTT — Method

Create a tensor train of a constant 'value' with spacial dimensions 'n' and tensor ranks 'r'.

TensorTrainApprox.constantTT — Method

Create a tensor train of a constant 'value' with spacial dimensions 'n' and tensor ranks 'r'.

TensorTrainApprox.createCTT — Method

Create a Chebyshev tensor train from a tensor train.

TensorTrainApprox.createCTT — Method

Create a Chebyshev tensor train from a tensor train.

TensorTrainApprox.createFTT — Method

Create a functional tensor train from a Chebyshev tensor train.

TensorTrainApprox.createFTT — Method

Create a functional tensor train from a Chebyshev tensor train.

TensorTrainApprox.createFTT — Method

Create a functional tensor train from a tensor train.

TensorTrainApprox.createFTT — Method

Create a functional tensor train from a tensor train.

TensorTrainApprox.createFTT — Method

Create a functional tensor train from a piecewise tensor train.

TensorTrainApprox.createPTT — Method

Create a piecewise linear tensor train from a tensor train.

TensorTrainApprox.decompress — Method

Recreates the dense array from a tensor train.

TensorTrainApprox.decompress_threaded — Method

Recreates the dense array from a tensor train.

TensorTrainApprox.frobenius — Method

Compute the Frobenius norm of a d-dimensional array.

TensorTrainApprox.ind2sub — Method

Find the array subindices from a linear index.

TensorTrainApprox.ind2sub — Method

Find the array subindices from a linear index.

TensorTrainApprox.randomTT — Method

Create a random tensor train with spacial dimensions 'n' and tensor ranks 'r'.

TensorTrainApprox.randomTT — Method

Create a random tensor train with spacial dimensions 'n' and tensor ranks 'r'.

TensorTrainApprox.times_cores — Method

Multiply two cores of a tensor train.

TensorTrainApprox.tsvd — Method

Compute the truncated SVD decomposition of a matrix.

TensorTrainApprox.frobenius — Method

Compute the Frobenius norm of a d-dimensional array.