TensorTrainApprox

Documentation for TensorTrainApprox.

TensorTrainApprox.DMRGcross

TensorTrainApprox.DMRGcross

TensorTrainApprox.DMRGcross

TensorTrainApprox.DMRGcross_generic

TensorTrainApprox.DMRGcross_generic

TensorTrainApprox.DMRGcross_generic

TensorTrainApprox.DMRGcross_generic_threaded

TensorTrainApprox.DMRGcross_generic_threaded

TensorTrainApprox.DMRGcross_generic_threaded

TensorTrainApprox.DMRGcross_threaded

TensorTrainApprox.DMRGcross_threaded

TensorTrainApprox.DMRGcross_threaded

TensorTrainApprox.TTCD_GC

TensorTrainApprox.TTCD_GH

TensorTrainApprox.TTCD_GL

TensorTrainApprox.TTCD_PL

TensorTrainApprox.TTChebyshev

TensorTrainApprox.TTLegendre

TensorTrainApprox.TTadd

TensorTrainApprox.TTconstant

TensorTrainApprox.TTconstant

TensorTrainApprox.TTderivative

TensorTrainApprox.TTderivative

TensorTrainApprox.TTevaluate

TensorTrainApprox.TTevaluate

TensorTrainApprox.TTevaluate

TensorTrainApprox.TTevaluate

TensorTrainApprox.TTgradient

TensorTrainApprox.TTgradient

TensorTrainApprox.TTintegrate_GC

TensorTrainApprox.TTintegrate_GC

TensorTrainApprox.TTintegrate_GH

TensorTrainApprox.TTintegrate_GH

TensorTrainApprox.TTintegrate_GL

TensorTrainApprox.TTintegrate_GL

TensorTrainApprox.TTintegrate_PL

TensorTrainApprox.TTinterp

TensorTrainApprox.TTmult

TensorTrainApprox.TTmult

TensorTrainApprox.TTmult

TensorTrainApprox.TTnorm

TensorTrainApprox.TTorthleft

TensorTrainApprox.TTorthleft

TensorTrainApprox.TTorthleftright

TensorTrainApprox.TTorthleftright

TensorTrainApprox.TTorthright

TensorTrainApprox.TTorthright

TensorTrainApprox.TTrandom

TensorTrainApprox.TTrandom

TensorTrainApprox.TTrounding

TensorTrainApprox.TTrounding

TensorTrainApprox.TTsubtract

TensorTrainApprox.TTsvd

TensorTrainApprox.bisection

TensorTrainApprox.chebyshev_evaluate

TensorTrainApprox.chebyshev_interp

TensorTrainApprox.chebyshev_nodes

TensorTrainApprox.chebyshev_polynomial

TensorTrainApprox.chebyshev_polynomial

TensorTrainApprox.chebyshev_polynomial_deriv

TensorTrainApprox.chebyshev_polynomial_deriv

TensorTrainApprox.chebyshev_weights

TensorTrainApprox.compute_marginal_GC

TensorTrainApprox.compute_marginal_GH

TensorTrainApprox.compute_marginal_GL

TensorTrainApprox.compute_marginal_PL

TensorTrainApprox.createCTT

TensorTrainApprox.createCTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createLTT

TensorTrainApprox.createLTT

TensorTrainApprox.createPTT

TensorTrainApprox.decompress

TensorTrainApprox.decompress_threaded

TensorTrainApprox.frobenius

TensorTrainApprox.frobenius

TensorTrainApprox.ind2sub

TensorTrainApprox.ind2sub

TensorTrainApprox.ind2sub

TensorTrainApprox.ind2sub

TensorTrainApprox.invert_cdf

TensorTrainApprox.invert_cdf

TensorTrainApprox.legendre_evaluate

TensorTrainApprox.legendre_interp

TensorTrainApprox.legendre_nodes

TensorTrainApprox.legendre_polynomial

TensorTrainApprox.legendre_polynomial

TensorTrainApprox.legendre_polynomial_deriv

TensorTrainApprox.legendre_polynomial_deriv

TensorTrainApprox.legendre_weights

TensorTrainApprox.normalize_node

TensorTrainApprox.normalize_node

TensorTrainApprox.rq

TensorTrainApprox.stack

TensorTrainApprox.times_cores

TensorTrainApprox.times_dim_1

TensorTrainApprox.times_dim_1

TensorTrainApprox.times_dim_3

TensorTrainApprox.times_dim_4

TensorTrainApprox.topK

TensorTrainApprox.trapazoidal

TensorTrainApprox.tsvd

TensorTrainApprox.DMRGcross — Method

Compute a tensor train approximation of a dense d-dimensional array. DMRG tensor compression based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross — Method

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross — Method

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_generic — Method

Compute a tensor train approximation of a dense d-dimensional array. DMRG tensor compression based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_generic — Method

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_generic — Method

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_generic_threaded — Method

Compute a tensor train approximation of a dense d-dimensional array. DMRG tensor compression based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_generic_threaded — Method

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_generic_threaded — Method

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_threaded — Method

Compute a tensor train approximation of a dense d-dimensional array. DMRG tensor compression based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_threaded — Method

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_threaded — Method

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.TTCD_GC — Method

Conditional distribution sampling when the nodes are Gauss-Chebyshev.

TensorTrainApprox.TTCD_GH — Method

Conditional distribution sampling when the nodes are Gauss-Hermite.

TensorTrainApprox.TTCD_GL — Method

Conditional distribution sampling when the nodes are Gauss-Legendre.

TensorTrainApprox.TTCD_PL — Method

Conditional distribution sampling when the nodes are uniformly spaced.

TensorTrainApprox.TTChebyshev — Method

Create a Chebyshev tensor train with ranks 'r', orders, 'order', and Chebyshev coefficients, 'θ'.

TensorTrainApprox.TTLegendre — Method

Create a Legendre tensor train with ranks 'r', orders, 'order', and Legendre coefficients, 'θ'.

TensorTrainApprox.TTadd — Method

Add two discrete tensor trains.

TensorTrainApprox.TTconstant — Method

Create a tensor train of a constant 'value' with spacial dimensions 'n' and tensor ranks 'r'.

TensorTrainApprox.TTconstant — Method

Create a tensor train of a constant 'value' with spacial dimensions 'n' and tensor ranks 'r'.

TensorTrainApprox.TTderivative — Method

Compute the derivative of a Chebyshev tensor train with respect to a variable.

TensorTrainApprox.TTderivative — Method

Compute the derivative of a Legendre tensor train with respect to a variable.

TensorTrainApprox.TTevaluate — Method

Evaluate a discrete tensor train at a vector of indexes of the dense array.

TensorTrainApprox.TTevaluate — Method

Evaluate a discrete tensor train at an index of the dense array.

TensorTrainApprox.TTevaluate — Method

Evaluate a functional tensor train at a vector of 'points'.

TensorTrainApprox.TTevaluate — Method

Evaluate a functional tensor train at 'point'.

TensorTrainApprox.TTgradient — Method

Compute the gradient vector of a Chebyshev tensor train.

TensorTrainApprox.TTgradient — Method

Compute the gradient vector of a Legendre tensor train.

TensorTrainApprox.TTintegrate_GC — Method

Integrate a discrete tensor train using Gauss-Chebyshev quadrature.

TensorTrainApprox.TTintegrate_GC — Method

Integrate a functional tensor train using Gauss-Chebyshev quadrature.

TensorTrainApprox.TTintegrate_GH — Method

Integrate a discrete tensor train using Gauss-Hermite quadrature.

TensorTrainApprox.TTintegrate_GH — Method

Integrate a functional tensor train using Gauss-Hermite quadrature.

TensorTrainApprox.TTintegrate_GL — Method

Integrate a discrete tensor train using Gauss-Legendre quadrature.

TensorTrainApprox.TTintegrate_GL — Method

Integrate a functional tensor train using Gauss-Legendre quadrature.

TensorTrainApprox.TTintegrate_PL — Method

Integrate a discrete tensor train using the trapazoidal method.

TensorTrainApprox.TTinterp — Method

Create an interpolating function from a functional tensor train.

TensorTrainApprox.TTmult — Method

Multiply a discrete tensor train and a scalar.

TensorTrainApprox.TTmult — Method

Multiply a discrete tensor train and a scalar.

TensorTrainApprox.TTmult — Method

Multiply a scalar and a discrete tensor train.

TensorTrainApprox.TTnorm — Method

Compute the norm of a discrete tensor train.

TensorTrainApprox.TTorthleft — Method

Left orthogonalise a discrete tensor train.

TensorTrainApprox.TTorthleft — Method

Left orthogonalise a discrete tensor train.

TensorTrainApprox.TTorthleftright — Method

Left orthogonalise a train up to core 'μ' and right orthogonalise the remaining cores.

TensorTrainApprox.TTorthleftright — Method

Left orthogonalise a train up to core 'μ' and right orthogonalise the remaining cores.

TensorTrainApprox.TTorthright — Method

Right orthogonalise a discrete tensor train.

TensorTrainApprox.TTorthright — Method

Right orthogonalise a discrete tensor train.

TensorTrainApprox.TTrandom — Method

Create a random tensor train with spacial dimensions 'n' and tensor ranks 'r'.

TensorTrainApprox.TTrandom — Method

Create a random tensor train with spacial dimensions 'n' and tensor ranks 'r'.

TensorTrainApprox.TTrounding — Method

Perform rounding on a discrete tensor train to find a more compact discrete tensor train representation.

TensorTrainApprox.TTrounding — Method

Perform rounding on a discrete tensor train to find a compact discrete tensor train representation.

TensorTrainApprox.TTsubtract — Method

Subtract two discrete tensor trains.

TensorTrainApprox.TTsvd — Method

Compute a tensor train approximation of a dense d-dimensional array. Tensor compression based on Oseledets and Tyrtyshnikov (2010).

TensorTrainApprox.bisection — Method

Use bisection to find the fix-point of the continuous function 'f'.

TensorTrainApprox.chebyshev_evaluate — Method

Evaluate a Chebyshev polynomial at 'point'.

TensorTrainApprox.chebyshev_interp — Method

Create a function that evaluates a Chebyshev polynomial at 'point'.

TensorTrainApprox.chebyshev_nodes — Method

Create a vector of 'n' Chebyshev nodes on 'domain' with element type 'T'.

TensorTrainApprox.chebyshev_polynomial — Method

Generate Chebyshev polynomials of 'order' at 'point'.

TensorTrainApprox.chebyshev_polynomial — Method

Generate Chebyshev polynomials of 'order' at each element of 'points'.

TensorTrainApprox.chebyshev_polynomial_deriv — Method

Construct the derivative of a Chebyshev polynomial at 'point'.

TensorTrainApprox.chebyshev_polynomial_deriv — Method

Construct the derivative of a Chebyshev polynomial at each element of 'points'.

TensorTrainApprox.chebyshev_weights — Method

Compute the weights for a univariate Chebyshev polynomial with 'order' on 'domain'.

TensorTrainApprox.compute_marginal_GC — Method

Integrate a discrete tensor train over all dimensions except 'ind' using Gauss-Chebyshev quadrature.

TensorTrainApprox.compute_marginal_GH — Method

Integrate a discrete tensor train over all dimensions except 'ind' using Gauss-Hermite quadrature.

TensorTrainApprox.compute_marginal_GL — Method

Integrate a discrete tensor train over all dimensions except 'ind' using Gauss-Legendre quadrature.

TensorTrainApprox.compute_marginal_PL — Method

Integrate a discrete tensor train over all dimensions except 'ind' using the trapazoidal method.

TensorTrainApprox.createCTT — Method

Create a Chebyshev tensor train from a discrete tensor train.

TensorTrainApprox.createCTT — Method

Create a Chebyshev tensor train from a discrete tensor train.

TensorTrainApprox.createFTT — Method

Create a functional tensor train from a Chebyshev tensor train.

TensorTrainApprox.createFTT — Method

Create a functional tensor train from a Legendre tensor train.

TensorTrainApprox.createFTT — Method

Create a functional tensor train from a discrete tensor train, assuming the nodes are Chebyshev roots.

TensorTrainApprox.createFTT — Method

Create a functional tensor train from a Chebyshev tensor train.

TensorTrainApprox.createFTT — Method

Create a functional tensor train from a Legendre tensor train.

TensorTrainApprox.createFTT — Method

Create a functional tensor train from a discrete tensor train, assuming the nodes are Chebyshev roots.

TensorTrainApprox.createFTT — Method

Create a functional tensor train from a piecewise tensor train.

TensorTrainApprox.createLTT — Method

Create a Legendre tensor train from a discrete tensor train.

TensorTrainApprox.createLTT — Method

Create a Legendre tensor train from a discrete tensor train.

TensorTrainApprox.createPTT — Method

Create a piecewise linear tensor train from a discrete tensor train.

TensorTrainApprox.decompress — Method

Recreates the dense array from a discrete tensor train.

TensorTrainApprox.decompress_threaded — Method

Recreates the dense array from a discrete tensor train.

TensorTrainApprox.frobenius — Method

Compute the Frobenius norm of a d-dimensional array.

TensorTrainApprox.ind2sub — Method

Find the array sub-indices from a linear index.

TensorTrainApprox.ind2sub — Method

Find the array sub-indices from a linear index.

TensorTrainApprox.ind2sub — Method

Find the array sub-indices from a linear index.

TensorTrainApprox.ind2sub — Method

Find the array sub-indices from a linear index.

TensorTrainApprox.invert_cdf — Method

Take a draw from a density with finite domain by inverting the CDF.

TensorTrainApprox.invert_cdf — Method

Take a draw from a density with unbounded domain by inverting the CDF.

TensorTrainApprox.legendre_evaluate — Method

Evaluate a Legendre polynomial at 'point'.

TensorTrainApprox.legendre_interp — Method

Create a function that evaluates a Legendre polynomial at 'point'.

TensorTrainApprox.legendre_nodes — Method

Create a vector of 'n' Legendre nodes on 'domain' with element type 'T'.

TensorTrainApprox.legendre_polynomial — Method

Generate Legendre polynomials of 'order' at 'point'.

TensorTrainApprox.legendre_polynomial — Method

Generate Legendre polynomials of 'order' at each element of 'points'.

TensorTrainApprox.legendre_polynomial_deriv — Method

Construct the derivative of a Legendre polynomial at 'point'.

TensorTrainApprox.legendre_polynomial_deriv — Method

Construct the derivative of a Legendre polynomial at each element of 'points'.

TensorTrainApprox.legendre_weights — Method

Compute the weights for a univariate Legendre polynomial with 'order' on 'domain'.

TensorTrainApprox.normalize_node — Method

Normalize 'node' so that it resides within the interval [-1,1].

TensorTrainApprox.normalize_node — Method

Normalize 'node' so that all elements reside within the interval [-1,1].

TensorTrainApprox.rq — Method

Perform the RQ matrix decomposition.

TensorTrainApprox.stack — Method

Horozontally concatenate two matrices.

TensorTrainApprox.times_cores — Method

Multiply two cores of a tensor train.

TensorTrainApprox.times_dim_1 — Method

Multiply a 3D array by a matrix along the first dimension of the 3D array.

TensorTrainApprox.times_dim_1 — Method

Multiply a 4D array by a matrix along the first dimension of the 4D array.

TensorTrainApprox.times_dim_3 — Method

Multiply a matrix by a 3D array along the third dimension of the 3D array.

TensorTrainApprox.times_dim_4 — Method

Multiply a matrix by a 4D array along the third dimension of the 4D array.

TensorTrainApprox.topK — Method

Find the 'K' rows of a matrix with the highest norm.

TensorTrainApprox.trapazoidal — Method

Use the trapazoidal method to integrate the cores of a discrete tensor train.

TensorTrainApprox.tsvd — Method

Compute the truncated SVD decomposition of a matrix with singular value threshold 'δ'.

TensorTrainApprox.frobenius — Method

Compute the Frobenius norm of a d-dimensional array.