TensorTrainApprox

Documentation for TensorTrainApprox.

TensorTrainApprox.DMRGcross

TensorTrainApprox.DMRGcross

TensorTrainApprox.DMRGcross

TensorTrainApprox.DMRGcross

TensorTrainApprox.DMRGcross_generic

TensorTrainApprox.DMRGcross_generic

TensorTrainApprox.DMRGcross_generic

TensorTrainApprox.DMRGcross_generic_threaded

TensorTrainApprox.DMRGcross_generic_threaded

TensorTrainApprox.DMRGcross_generic_threaded

TensorTrainApprox.DMRGcross_threaded

TensorTrainApprox.DMRGcross_threaded

TensorTrainApprox.DMRGcross_threaded

TensorTrainApprox.DMRGcross_threaded

TensorTrainApprox.TTCD_GC

TensorTrainApprox.TTCD_GH

TensorTrainApprox.TTCD_GL

TensorTrainApprox.TTCD_PL

TensorTrainApprox.TTChebyshev

TensorTrainApprox.TTLegendre

TensorTrainApprox.TTOpt

TensorTrainApprox.TTSIRT_GC

TensorTrainApprox.TTSIRT_GH

TensorTrainApprox.TTSIRT_GL

TensorTrainApprox.TTSIRT_PL

TensorTrainApprox.TTadd

TensorTrainApprox.TTcompute_marginal_GC

TensorTrainApprox.TTcompute_marginal_GH

TensorTrainApprox.TTcompute_marginal_GL

TensorTrainApprox.TTcompute_marginal_PL

TensorTrainApprox.TTconstant

TensorTrainApprox.TTconstant

TensorTrainApprox.TTderivative

TensorTrainApprox.TTderivative

TensorTrainApprox.TTevaluate

TensorTrainApprox.TTevaluate

TensorTrainApprox.TTevaluate

TensorTrainApprox.TTevaluate

TensorTrainApprox.TTexp

TensorTrainApprox.TTextremize

TensorTrainApprox.TTextremize

TensorTrainApprox.TTextremize

TensorTrainApprox.TTgradient

TensorTrainApprox.TTgradient

TensorTrainApprox.TTinner_prod

TensorTrainApprox.TTintegrate_GC

TensorTrainApprox.TTintegrate_GC

TensorTrainApprox.TTintegrate_GH

TensorTrainApprox.TTintegrate_GH

TensorTrainApprox.TTintegrate_GL

TensorTrainApprox.TTintegrate_LC

TensorTrainApprox.TTintegrate_PL

TensorTrainApprox.TTinterp

TensorTrainApprox.TTmult

TensorTrainApprox.TTmult

TensorTrainApprox.TTmult

TensorTrainApprox.TTnorm

TensorTrainApprox.TToptimize

TensorTrainApprox.TToptimize

TensorTrainApprox.TToptimize

TensorTrainApprox.TTorthleft

TensorTrainApprox.TTorthleft

TensorTrainApprox.TTorthleftright

TensorTrainApprox.TTorthleftright

TensorTrainApprox.TTorthright

TensorTrainApprox.TTorthright

TensorTrainApprox.TTpower

TensorTrainApprox.TTrandom

TensorTrainApprox.TTrandom

TensorTrainApprox.TTrounding

TensorTrainApprox.TTrounding

TensorTrainApprox.TTsize

TensorTrainApprox.TTsquared

TensorTrainApprox.TTsubtract

TensorTrainApprox.TTsvd

TensorTrainApprox.TTsvd

TensorTrainApprox.bisection

TensorTrainApprox.cheb_evaluate

TensorTrainApprox.cheb_nodes

TensorTrainApprox.cheb_polynomial

TensorTrainApprox.cheb_polynomial

TensorTrainApprox.cheb_polynomial_deriv

TensorTrainApprox.cheb_polynomial_deriv

TensorTrainApprox.cheb_weights

TensorTrainApprox.chebyshev_interp

TensorTrainApprox.createCTT

TensorTrainApprox.createCTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createLTT

TensorTrainApprox.createLTT

TensorTrainApprox.createPTT

TensorTrainApprox.decompress

TensorTrainApprox.decompress_threaded

TensorTrainApprox.frobenius

TensorTrainApprox.frobenius

TensorTrainApprox.ind2sub

TensorTrainApprox.ind2sub

TensorTrainApprox.invert_cdf

TensorTrainApprox.invert_cdf

TensorTrainApprox.legendre_evaluate

TensorTrainApprox.legendre_interp

TensorTrainApprox.legendre_nodes

TensorTrainApprox.legendre_polynomial

TensorTrainApprox.legendre_polynomial

TensorTrainApprox.legendre_polynomial_deriv

TensorTrainApprox.legendre_polynomial_deriv

TensorTrainApprox.legendre_weights

TensorTrainApprox.normalize_node

TensorTrainApprox.normalize_node

TensorTrainApprox.rq

TensorTrainApprox.stack

TensorTrainApprox.times_cores

TensorTrainApprox.times_dim_1

TensorTrainApprox.times_dim_1

TensorTrainApprox.times_dim_3

TensorTrainApprox.times_dim_4

TensorTrainApprox.topK

TensorTrainApprox.trapazoidal

TensorTrainApprox.tsvd

TensorTrainApprox.tsvd

TensorTrainApprox.DMRGcross — Method

t = DMRGcross(b,μ,tol,maxsweeps) t = DMRGcross(b,μ,tol)

Compute a tensor train approximation of a dense d-dimensional array. DMRG tensor compression based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross — Method

t = DMRGcross(f,nodes,μ,tol,initial,maxsweeps) t = DMRGcross(f,nodes,μ,tol,initial)

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross — Method

t = DMRGcross(f,nodes,μ,tol,maxsweeps) t = DMRGcross(f,nodes,μ,tol)

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross — Method

t = DMRGcross(f,nodes,initial)

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020) with the ranks imposed.

TensorTrainApprox.DMRGcross_generic — Method

t = DMRGcrossgeneric(b,μ,tol,maxsweeps) t = DMRGcrossgeneric(b,μ,tol)

Compute a tensor train approximation of a dense d-dimensional array. DMRG tensor compression based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_generic — Method

t = DMRGcrossgeneric(f,nodes,μ,tol,initial,maxsweeps) t = DMRGcrossgeneric(f,nodes,μ,tol,initial)

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_generic — Method

t = DMRGcrossgeneric(f,nodes,μ,tol,maxsweeps) t = DMRGcrossgeneric(f,nodes,μ,tol)

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_generic_threaded — Method

t = DMRGcrossgenericthreaded(b,μ,tol,maxsweeps) t = DMRGcrossgenericthreaded(b,μ,tol)

Compute a tensor train approximation of a dense d-dimensional array. DMRG tensor compression based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_generic_threaded — Method

t = DMRGcrossgenericthreaded(f,nodes,μ,tol,initial,maxsweeps) t = DMRGcrossgenericthreaded(f,nodes,μ,tol,initial)

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_generic_threaded — Method

t = DMRGcrossgenericthreaded(f,nodes,μ,tol,maxsweeps) t = DMRGcrossgenericthreaded(f,nodes,μ,tol)

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_threaded — Method

t = DMRGcrossthreaded(b,μ,tol,maxsweeps) t = DMRGcrossthreaded(b,μ,tol)

Compute a tensor train approximation of a dense d-dimensional array. DMRG tensor compression based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_threaded — Method

t = DMRGcrossthreaded(f,nodes,μ,tol,initial,maxsweeps) t = DMRGcrossthreaded(f,nodes,μ,tol,initial)

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_threaded — Method

t = DMRGcrossthreaded(f,nodes,μ,tol,maxsweeps) t = DMRGcrossthreaded(f,nodes,μ,tol)

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_threaded — Method

t = DMRGcross_threaded(f,nodes,initial)

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.TTCD_GC — Method

sample = TTCDGC(train,N,domain,seed) sample = TTCDGC(train,N,domain)

Conditional distribution sampling when the nodes are Gauss-Chebyshev.

TensorTrainApprox.TTCD_GH — Method

sample = TTCDGH(train,N,seed) sample = TTCDGH(train,N)

Conditional distribution sampling when the nodes are Gauss-Hermite.

TensorTrainApprox.TTCD_GL — Method

sample = TTCDGL(train,N,domain,seed) sample = TTCDGL(train,N,domain)

Conditional distribution sampling when the nodes are Gauss-Legendre.

TensorTrainApprox.TTCD_PL — Method

sample = TTCDPL(train,N,domain,seed) sample = TTCDPL(train,N,domain)

Conditional distribution sampling when the nodes are uniformly spaced.

TensorTrainApprox.TTChebyshev — Method

t = TTChebyshev(r,order,θ)

Create a Chebyshev tensor train with ranks, 'r', orders, 'order', and Chebyshev coefficients, 'θ'.

TensorTrainApprox.TTLegendre — Method

t = TTLegendre(r,order,θ)

Create a Legendre tensor train with ranks, 'r', orders, 'order', and Legendre coefficients, 'θ'.

TensorTrainApprox.TTOpt — Method

soln = TTOpt(f,nodes,rmax,sweeps)

Minimize a function by discretizing and compressing it using a tensor train with maximum rank, rmax.

TensorTrainApprox.TTSIRT_GC — Method

sample = TTSIRTGC(train,N,domain,seed) sample = TTSIRTGC(train,N,domain)

Conditional distribution sampling from the square of a tensor train when the nodes are Gauss-Chebyshev.

TensorTrainApprox.TTSIRT_GH — Method

sample = TTSIRTGH(train,N,seed) sample = TTSIRTGH(train,N)

Conditional distribution sampling from the square of a tensor train when the nodes are Gauss-Hermite.

TensorTrainApprox.TTSIRT_GL — Method

sample = TTSIRTGL(train,N,domain,seed) sample = TTSIRTGL(train,N,domain)

Conditional distribution sampling from the square of a tensor train when the nodes are Gauss-Legendre.

TensorTrainApprox.TTSIRT_PL — Method

sample = TTSIRTPL(train,N,domain,seed) sample = TTSIRTPL(train,N,domain)

Conditional distribution sampling from the square of a tensor train when the nodes are uniformly spaced.

TensorTrainApprox.TTadd — Function

t = TTadd(traina,trainb,anchor)

Add two discrete tensor trains. When the trains have different lengths, anchor determines whether the trains align from the first core or the last core.

TensorTrainApprox.TTcompute_marginal_GC — Method

margin = TTcomputeintegrateGC(train,ind,domain)

Integrate a discrete tensor train over all dimensions except 'ind' using Gauss-Chebyshev quadrature.

TensorTrainApprox.TTcompute_marginal_GH — Method

margin = TTcomputeintegrateGH(train,ind)

Integrate a discrete tensor train over all dimensions except 'ind' using Gauss-Hermite quadrature.

TensorTrainApprox.TTcompute_marginal_GL — Method

margin = TTcomputeintegrateGL(train,ind,domain)

Integrate a discrete tensor train over all dimensions except 'ind' using Gauss-Legendre quadrature.

TensorTrainApprox.TTcompute_marginal_PL — Method

margin = TTcomputeintegratePL(train,ind,domain)

Integrate a discrete tensor train over all dimensions except 'ind' using the trapazoidal method.

TensorTrainApprox.TTconstant — Method

t = TTconstant(val,n,r)

Create a tensor train of a constant 'value' with spacial dimensions 'n' and tensor ranks 'r'.

TensorTrainApprox.TTconstant — Method

t = TTconstant(val,n,r)

Create a tensor train of a constant 'value' with spacial dimensions 'n' and tensor ranks 'r'.

TensorTrainApprox.TTderivative — Method

deriv = TTderivative(train,x,pos,domain)

Compute the derivative of a Chebyshev tensor train with respect to a variable.

TensorTrainApprox.TTderivative — Method

deriv = TTderivative(train,x,pos,domain)

Compute the derivative of a Legendre tensor train with respect to a variable.

TensorTrainApprox.TTevaluate — Method

yhat = TTevaluate(train,index)

Evaluate a discrete tensor train at a vector of indexes of the dense array.

TensorTrainApprox.TTevaluate — Method

yhat = TTevaluate(train,index)

Evaluate a discrete tensor train at an index of the dense array.

TensorTrainApprox.TTevaluate — Method

yhat = TTevaluate(train,points)

Evaluate a functional tensor train at a vector of 'points'.

TensorTrainApprox.TTevaluate — Method

yhat = TTevaluate(train,point)

Evaluate a functional tensor train at 'point'.

TensorTrainApprox.TTexp — Method

t = TTexp(train,N,tol)

Approximate the exponetial of a tensor train in tensor train format

TensorTrainApprox.TTextremize — Method

soln = TTextremize(train,K,state)

Extremize over the final (d-ds) cores of a tensor train, evaluating the first ds cores at the indices given in 'state'. The output could be a maxima or a minima.

TensorTrainApprox.TTextremize — Method

soln = TTextremize(train,K)

Extremize over the cores of a tensor train. The output could be a maxima or a minima.

TensorTrainApprox.TTextremize — Method

soln = TTextremize(train,K,nodes,state)

Extremize over the final (d-ds) cores of a tensor train, evaluating the first ds cores at the point given in 'state'. The output could be a maxima or a minima.

TensorTrainApprox.TTgradient — Method

grad = TTgradient(train,x,domain)

Compute the gradient vector of a Chebyshev tensor train.

TensorTrainApprox.TTgradient — Method

grad = TTgradient(train,x,domain)

Compute the gradient vector of a Legendre tensor train.

TensorTrainApprox.TTinner_prod — Method

innerprod = TTinnerprod(traina,trainb)

Compute the inner product of two tensor trains.

TensorTrainApprox.TTintegrate_GC — Method

area = TTintegrate_GC(train,domain)

Integrate a discrete tensor train using Gauss-Chebyshev quadrature.

TensorTrainApprox.TTintegrate_GC — Method

area = TTintegrate_GC(train,nodes,domain)

Integrate a functional tensor train using Gauss-Chebyshev quadrature.

TensorTrainApprox.TTintegrate_GH — Method

area = TTintegrate_GH(train)

Integrate a discrete tensor train using Gauss-Hermite quadrature.

TensorTrainApprox.TTintegrate_GH — Method

area = TTintegrate_GH(train,nodes)

Integrate a functional tensor train using Gauss-Hermite quadrature.

TensorTrainApprox.TTintegrate_GL — Method

area = TTintegrate_GL(train,domain)

Integrate a discrete tensor train using Gauss-Legendre quadrature.

TensorTrainApprox.TTintegrate_LC — Method

area = TTintegrate_GL(train,nodes,domain)

Integrate a functional tensor train using Gauss-Legendre quadrature.

TensorTrainApprox.TTintegrate_PL — Method

area = TTintegrate_PL(train,domain)

Integrate a discrete tensor train using the trapazoidal method.

TensorTrainApprox.TTinterp — Method

f = TTinterp(train)

Create an interpolating function from a functional tensor train.

TensorTrainApprox.TTmult — Method

t = TTmult(train,s)

Multiply a discrete tensor train and a scalar.

TensorTrainApprox.TTmult — Method

t = TTmult(train,s)

Multiply a discrete tensor train and a scalar.

TensorTrainApprox.TTmult — Method

t = TTmult(s,train)

Multiply a scalar and a discrete tensor train.

TensorTrainApprox.TTnorm — Method

TTnorm(train)

Compute the norm of a discrete tensor train.

TensorTrainApprox.TToptimize — Method

soln = TToptimize(train,K,state)

Optimize over the cores final (d-ds) of a tensor train, evaluating the first ds cores at the indices given in 'state' and returning the maximum and the minimum.

TensorTrainApprox.TToptimize — Method

soln = TToptimize(train,K)

Optimize over the cores of a tensor train, returning the maximum and the minimum.

TensorTrainApprox.TToptimize — Method

soln = TToptimize(train,K,nodes,state)

Optimize over the cores final (d-ds) of a tensor train, evaluating the first ds cores at the point given in 'state' and returning the maximum and the minimum.

TensorTrainApprox.TTorthleft — Method

t = TTorthleft(train)

Left orthogonalise a discrete tensor train.

TensorTrainApprox.TTorthleft — Method

t = TTorthleft(train)

Left orthogonalise a discrete tensor train.

TensorTrainApprox.TTorthleftright — Method

t = TTorthleftright(train,μ)

Left orthogonalise a train up to core 'μ' and right orthogonalise the remaining cores.

TensorTrainApprox.TTorthleftright — Method

t = TTorthleftright(train,μ)

Left orthogonalise a train up to core 'μ' and right orthogonalise the remaining cores.

TensorTrainApprox.TTorthright — Method

t = TTorthright(train)

Right orthogonalise a discrete tensor train.

TensorTrainApprox.TTorthright — Method

t = TTorthright(train)

Right orthogonalise a discrete tensor train.

TensorTrainApprox.TTpower — Method

t = TTpower(train,N)

Construct the N'th power of a tensor train.

TensorTrainApprox.TTrandom — Method

t = TTrandom(n,r,T)

Create a random tensor train with spacial dimensions 'n' and tensor ranks 'r' with eltype(T).

TensorTrainApprox.TTrandom — Method

t = TTrandom(n,r,T)

Create a random tensor train with spacial dimensions 'n' and tensor ranks 'r' with eltype(T).

TensorTrainApprox.TTrounding — Method

t = TTrounding(train,tol)

Perform rounding on a discrete tensor train to find a more compact discrete tensor train representation.

TensorTrainApprox.TTrounding — Method

t = TTrounding(train,tol)

Perform rounding on a discrete tensor train to find a compact discrete tensor train representation.

TensorTrainApprox.TTsize — Method

n = TTsize(train)

Compute the (implied) size of a decompressed tensor train.

TensorTrainApprox.TTsquared — Method

t = TTsquard(train)

Construct the square of a tensor train.

TensorTrainApprox.TTsubtract — Function

t = TTsubstract(traina,trainb,anchor)

Subtract two discrete tensor trains. When the trains have different lengths, anchor determines whether the trains align from the first core or the last core.

TensorTrainApprox.TTsvd — Method

t = TTsvd(a,tol)

Tensor compression based on Oseledets and Tyrtyshnikov (2010).

TensorTrainApprox.TTsvd — Method

t = TTsvd(a,r)

Tensor compression based on Oseledets and Tyrtyshnikov (2010) with imposed ranks, 'r'.

TensorTrainApprox.bisection — Method

Use bisection to find the fix-point of the continuous function 'f'.

TensorTrainApprox.cheb_evaluate — Method

yhat = cheb_evaluate(w,point,order,domain)

Evaluate a Chebyshev polynomial at 'point'.

TensorTrainApprox.cheb_nodes — Method

nodes = cheb_nodes(n,domain)

Create a vector of 'n' Chebyshev nodes on 'domain' with element type 'T'.

TensorTrainApprox.cheb_polynomial — Method

P = chebyshev_polynmial(order,point)

Generate Chebyshev polynomials of 'order' at 'point'.

TensorTrainApprox.cheb_polynomial — Method

P = chebyshev_polynmial(order,points)

Generate Chebyshev polynomials of 'order' at each element of 'points'.

TensorTrainApprox.cheb_polynomial_deriv — Method

deriv = chebpolynomialderiv(order,point)

Construct the derivative of a Chebyshev polynomial at 'point'.

TensorTrainApprox.cheb_polynomial_deriv — Method

deriv = chebpolynomialderiv(order,points)

Construct the derivative of a Chebyshev polynomial at each element of 'points'.

TensorTrainApprox.cheb_weights — Method

w = cheb_weights(y,nodes,order,domain)

Compute the weights for a univariate Chebyshev polynomial with 'order' on 'domain'.

TensorTrainApprox.chebyshev_interp — Method

f = chebyshev_interp(y,nodes,order,domain)

Create a function that evaluates a Chebyshev polynomial at 'point'.

TensorTrainApprox.createCTT — Method

trainCTT = createCTT(g,nodes,domain)

Create a Chebyshev tensor train from a discrete tensor train.

TensorTrainApprox.createCTT — Method

trainCTT = createCTT(g,order,nodes,domain)

Create a Chebyshev tensor train from a discrete tensor train.

TensorTrainApprox.createFTT — Method

trainFTT = createFTT(g,domain)

Create a functional tensor train from a Chebyshev tensor train.

TensorTrainApprox.createFTT — Method

trainFTT = createFTT(g,domain)

Create a functional tensor train from a Legendre tensor train.

TensorTrainApprox.createFTT — Method

trainFTT = createFTT(g,nodes,domain)

Create a functional tensor train from a discrete tensor train, assuming the nodes are Chebyshev roots.

TensorTrainApprox.createFTT — Method

trainFTT = createFTT(g,order,domain)

Create a functional tensor train from a Chebyshev tensor train.

TensorTrainApprox.createFTT — Method

trainFTT = createFTT(g,order,domain)

Create a functional tensor train from a Legendre tensor train.

TensorTrainApprox.createFTT — Method

trainFTT = createFTT(g,order,nodes,domain)

Create a functional tensor train from a discrete tensor train, assuming the nodes are Chebyshev roots.

TensorTrainApprox.createFTT — Method

trainFTT = createFTT(g,nodes)

Create a functional tensor train from a piecewise tensor train.

TensorTrainApprox.createLTT — Method

trainLTT = createLTT(g,nodes,domain)

Create a Legendre tensor train from a discrete tensor train.

TensorTrainApprox.createLTT — Method

trainLTT = createLTT(g,order,nodes,domain)

Create a Legendre tensor train from a discrete tensor train.

TensorTrainApprox.createPTT — Method

trainPTT = createPTT(g)

Create a piecewise linear tensor train from a discrete tensor train.

TensorTrainApprox.decompress — Method

a = decompress(train)

Recreates the dense array from a discrete tensor train.

TensorTrainApprox.decompress_threaded — Method

a = decompress_threaded(train)

Recreates the dense array from a discrete tensor train.

TensorTrainApprox.frobenius — Method

t = frobenius(M)

Compute the Frobenius norm of a d-dimensional array.

TensorTrainApprox.ind2sub — Method

s = ind2sub(i,dims)

Find the array sub-indices from a linear index.

TensorTrainApprox.ind2sub — Method

s = ind2sub(i,dims)

Find the array sub-indices from a vector of linear indexes.

TensorTrainApprox.invert_cdf — Method

Take a draw from a density with finite domain by inverting the CDF.

TensorTrainApprox.invert_cdf — Method

Take a draw from a density with unbounded domain by inverting the CDF.

TensorTrainApprox.legendre_evaluate — Method

yhat = legendre_evaluate(w,point,order,domain)

Evaluate a Legendre polynomial at 'point'.

TensorTrainApprox.legendre_interp — Method

f = legendre_interp(y,nodes,order,domain)

Create a function that evaluates a Legendre polynomial at 'point'.

TensorTrainApprox.legendre_nodes — Method

nodes= legendre_nodes(n,domain)

Create a vector of 'n' Legendre nodes on 'domain' with element type 'T'.

TensorTrainApprox.legendre_polynomial — Method

P = legendre_polynomial(order,point)

Generate Legendre polynomials of 'order' at 'point'.

TensorTrainApprox.legendre_polynomial — Method

P = legendre_polynomial(order,points)

Generate Legendre polynomials of 'order' at each element of 'points'.

TensorTrainApprox.legendre_polynomial_deriv — Method

deriv = legendrepolynomialderiv(order,point)

Construct the derivative of a Legendre polynomial at 'point'.

TensorTrainApprox.legendre_polynomial_deriv — Method

deriv = legendrepolynomialderiv(order,points)

Construct the derivative of a Legendre polynomial at each element of 'points'.

TensorTrainApprox.legendre_weights — Method

w = legendre_weights(y,nodes,order,domain)

Compute the weights for a univariate Legendre polynomial with 'order' on 'domain'.

TensorTrainApprox.normalize_node — Method

normalize_node(node,domain)

Normalize 'node' so that it resides within the interval [-1,1].

TensorTrainApprox.normalize_node — Method

normalize_node(node,domain)

Normalize 'node' so that all elements reside within the interval [-1,1].

TensorTrainApprox.rq — Method

R,Q = rq(A)

Perform the RQ matrix decomposition.

TensorTrainApprox.stack — Method

Horozontally concatenate two matrices.

TensorTrainApprox.times_cores — Method

t = times_cores(a,b)

Multiply two cores of a tensor train.

TensorTrainApprox.times_dim_1 — Method

t = timesdim1(a,b)

Multiply a 3D array (b) by a matrix (a) along the first dimension of the 3D array.

TensorTrainApprox.times_dim_1 — Method

t = timesdim1(a,b)

Multiply a 4D array (b) by a matrix (a) along the first dimension of the 4D array.

TensorTrainApprox.times_dim_3 — Method

t = timesdim3(a,b)

Multiply a matrix (b) by a 3D array (a) along the third dimension of the 3D array.

TensorTrainApprox.times_dim_4 — Method

t = timesdim4(a,b)

Multiply a matrix (b) by a 4D array (a) along the third dimension of the 4D array.

TensorTrainApprox.topK — Method

Find the 'K' rows of a matrix with the highest norm.

TensorTrainApprox.trapazoidal — Method

Use the trapazoidal method to integrate the cores of a discrete tensor train.

TensorTrainApprox.tsvd — Method

t = tsvd(M,r)

Compute the truncated SVD decomposition of a matrix with imposed rank 'r'.

TensorTrainApprox.tsvd — Method

t = tsvd(M,δ)

Compute the truncated SVD decomposition of a matrix with singular value threshold 'δ'.

TensorTrainApprox.frobenius — Method

t = frobenius(M)

Compute the Frobenius norm of a d-dimensional array.