TensorTrainTools

Documentation for TensorTrainTools.

TensorTrainTools.DMRGcross

TensorTrainTools.DMRGcross

TensorTrainTools.DMRGcross_generic

TensorTrainTools.DMRGcross_generic

TensorTrainTools.DMRGcross_generic_threaded

TensorTrainTools.DMRGcross_generic_threaded

TensorTrainTools.DMRGcross_threaded

TensorTrainTools.DMRGcross_threaded

TensorTrainTools.TTCD_GC

TensorTrainTools.TTCD_GH

TensorTrainTools.TTCD_GL

TensorTrainTools.TTCD_PL

TensorTrainTools.TTChebyshev

TensorTrainTools.TTHadamard

TensorTrainTools.TTLegendre

TensorTrainTools.TTMH

TensorTrainTools.TTMHlog

TensorTrainTools.TTOpt

TensorTrainTools.TTOpt

TensorTrainTools.TTSIRT_GC

TensorTrainTools.TTSIRT_GH

TensorTrainTools.TTSIRT_GL

TensorTrainTools.TTSIRT_PL

TensorTrainTools.TTadd

TensorTrainTools.TTals

TensorTrainTools.TTals

TensorTrainTools.TTcompute_marginal_GC

TensorTrainTools.TTcompute_marginal_GH

TensorTrainTools.TTcompute_marginal_GL

TensorTrainTools.TTcompute_marginal_PL

TensorTrainTools.TTconstant

TensorTrainTools.TTcreateCTT

TensorTrainTools.TTcreateFTT

TensorTrainTools.TTcreateLTT

TensorTrainTools.TTcreatePTT

TensorTrainTools.TTdecompress

TensorTrainTools.TTdecompress_threaded

TensorTrainTools.TTderivative

TensorTrainTools.TTdescent

TensorTrainTools.TTdescent

TensorTrainTools.TTdescent_step

TensorTrainTools.TTevaluate

TensorTrainTools.TTevaluate

TensorTrainTools.TTexp

TensorTrainTools.TTextremize

TensorTrainTools.TTgradient

TensorTrainTools.TThessian

TensorTrainTools.TTinner_prod

TensorTrainTools.TTinner_prod2

TensorTrainTools.TTintegrate_GC

TensorTrainTools.TTintegrate_GC

TensorTrainTools.TTintegrate_GH

TensorTrainTools.TTintegrate_GH

TensorTrainTools.TTintegrate_GL

TensorTrainTools.TTintegrate_LC

TensorTrainTools.TTintegrate_PL

TensorTrainTools.TTinterp

TensorTrainTools.TTmult

TensorTrainTools.TTnewton

TensorTrainTools.TTnewton

TensorTrainTools.TTnewton_step

TensorTrainTools.TTnorm

TensorTrainTools.TTones

TensorTrainTools.TToptimize

TensorTrainTools.TTorthleft

TensorTrainTools.TTorthleftright

TensorTrainTools.TTorthright

TensorTrainTools.TTpower

TensorTrainTools.TTrand

TensorTrainTools.TTrandn

TensorTrainTools.TTreverse_indices

TensorTrainTools.TTrounding

TensorTrainTools.TTsize

TensorTrainTools.TTsquared

TensorTrainTools.TTsubtract

TensorTrainTools.TTsvd

TensorTrainTools.TTsvd

TensorTrainTools.TTsvd_threaded

TensorTrainTools.TTzeros

TensorTrainTools.bisection

TensorTrainTools.bracket_nodes

TensorTrainTools.cheb_evaluate

TensorTrainTools.cheb_nodes

TensorTrainTools.cheb_polynomial

TensorTrainTools.cheb_polynomial_deriv

TensorTrainTools.cheb_weights

TensorTrainTools.chebyshev_interp

TensorTrainTools.ind2sub

TensorTrainTools.invert_cdf

TensorTrainTools.legendre_evaluate

TensorTrainTools.legendre_interp

TensorTrainTools.legendre_nodes

TensorTrainTools.legendre_polynomial

TensorTrainTools.legendre_polynomial_deriv

TensorTrainTools.legendre_polynomial_sec_deriv

TensorTrainTools.legendre_weights

TensorTrainTools.normalize_node

TensorTrainTools.piecewise_linear_evaluate

TensorTrainTools.piecewise_linear_nodes

TensorTrainTools.rq

TensorTrainTools.stack

TensorTrainTools.times_cores

TensorTrainTools.times_dim_1

TensorTrainTools.times_dim_1

TensorTrainTools.times_dim_3

TensorTrainTools.times_dim_4

TensorTrainTools.topK

TensorTrainTools.trapazoidal

TensorTrainTools.tsvd

TensorTrainTools.DMRGcross — Method

Compute a tensor train approximation of a dense d-dimensional array, B, with max-volume factor, μ > 1.0.

Signatures

t = DMRGcross(B,μ,tol) t = DMRGcross(B,μ,tol,maxsweeps) t = DMRGcross(B,μ,r) t = DMRGcross(B,μ,r,maxsweeps)

TensorTrainTools.DMRGcross — Method

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array.

Signatures

t = DMRGcross(f,nodes,μ,tol) t = DMRGcross(f,nodes,μ,tol,maxsweeps) t = DMRGcross(f,nodes,μ,tol,initial) t = DMRGcross(f,nodes,μ,tol,initial,maxsweeps) t = DMRGcross(f,nodes,μ,r) t = DMRGcross(f,nodes,μ,r,maxsweeps)

TensorTrainTools.DMRGcross_generic — Method

Compute a tensor train approximation of a dense d-dimensional array, B, with max-volume factor, μ > 1.0.

Signatures

t = DMRGcrossgeneric(B,μ,tol) t = DMRGcrossgeneric(B,μ,tol,maxsweeps) t = DMRGcrossgeneric(B,μ,r) t = DMRGcrossgeneric(B,μ,r,maxsweeps)

TensorTrainTools.DMRGcross_generic — Method

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array.

Signatures

t = DMRGcrossgeneric(f,nodes,μ,tol) t = DMRGcrossgeneric(f,nodes,μ,tol,maxsweeps) t = DMRGcrossgeneric(f,nodes,μ,tol,initial) t = DMRGcrossgeneric(f,nodes,μ,tol,initial,maxsweeps) t = DMRGcrossgeneric(f,nodes,μ,r) t = DMRGcrossgeneric(f,nodes,μ,r,maxsweeps)

TensorTrainTools.DMRGcross_generic_threaded — Method

Compute a tensor train approximation of a dense d-dimensional array, B, with max-volume factor, μ > 1.0.

Signatures

t = DMRGcrossgenericthreaded(B,μ,tol) t = DMRGcrossgenericthreaded(B,μ,tol,maxsweeps) t = DMRGcrossgenericthreaded(B,μ,r) t = DMRGcrossgenericthreaded(B,μ,r,maxsweeps)

TensorTrainTools.DMRGcross_generic_threaded — Method

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array.

Signatures

t = DMRGcrossgenericthreaded(f,nodes,μ,tol) t = DMRGcrossgenericthreaded(f,nodes,μ,tol,maxsweeps) t = DMRGcrossgenericthreaded(f,nodes,μ,tol,initial) t = DMRGcrossgenericthreaded(f,nodes,μ,tol,initial,maxsweeps) t = DMRGcrossgenericthreaded(f,nodes,μ,r) t = DMRGcrossgenericthreaded(f,nodes,μ,r,maxsweeps)

TensorTrainTools.DMRGcross_threaded — Method

Compute a tensor train approximation of a dense d-dimensional array, B, with max-volume factor, μ > 1.0.

Signatures

t = DMRGcrossthreaded(B,μ,tol) t = DMRGcrossthreaded(B,μ,tol,maxsweeps) t = DMRGcrossthreaded(B,μ,r) t = DMRGcrossthreaded(B,μ,r,maxsweeps)

TensorTrainTools.DMRGcross_threaded — Method

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array.

Signatures

t = DMRGcrossthreaded(f,nodes,μ,tol) t = DMRGcrossthreaded(f,nodes,μ,tol,maxsweeps) t = DMRGcrossthreaded(f,nodes,μ,tol,initial) t = DMRGcrossthreaded(f,nodes,μ,tol,initial,maxsweeps) t = DMRGcrossthreaded(f,nodes,μ,r) t = DMRGcrossthreaded(f,nodes,μ,r,maxsweeps)

TensorTrainTools.TTCD_GC — Method

Conditional distribution sampling when the nodes are Gauss-Chebyshev.

Signatures

sample = TTCDGC(train,N,domain) sample = TTCDGC(train,N,domain,seed)

TensorTrainTools.TTCD_GH — Method

Conditional distribution sampling when the nodes are Gauss-Hermite.

Signatures

sample = TTCDGH(train,N) sample = TTCDGH(train,N,seed)

TensorTrainTools.TTCD_GL — Method

Conditional distribution sampling when the nodes are Gauss-Legendre.

Signatures

sample = TTCDGL(train,N,domain) sample = TTCDGL(train,N,domain,seed)

TensorTrainTools.TTCD_PL — Method

Conditional distribution sampling when the nodes are uniformly spaced.

Signatures

sample = TTCDPL(train,N,domain) sample = TTCDPL(train,N,domain,seed)

TensorTrainTools.TTChebyshev — Method

Create a Chebyshev tensor train with ranks, 'r', orders, 'order', and Chebyshev coefficients, 'θ'.

Signature

t = TTChebyshev(r,order,θ)

TensorTrainTools.TTHadamard — Method

Take the Hadamard product of two tensor trains.

Signature

t = TTHadamard(train1,train2)

TensorTrainTools.TTLegendre — Method

Create a Legendre tensor train with ranks, 'r', orders, 'order', and Legendre coefficients, 'θ'.

Signature

t = TTLegendre(r,order,θ)

TensorTrainTools.TTMH — Method

Metropolis-Hasting correction of a sample based on the density and its functional tensor train approximation.

Signatures

sample = TTMH(pdf,train,sample) sample = TTMH(pdf,train,sample,seed)

TensorTrainTools.TTMHlog — Method

Metropolis-Hasting correction of a sample based on the log-density and its functional tensor train approximation.

Signatures

sample = TTMHlog(logpdf,logtrain,sample) sample = TTMHlog(logpdf,logtrain,sample,seed)

TensorTrainTools.TTOpt — Method

Minimize a function by discretizing and compressing it using a tensor train with maximum rank, rmax.

Signature

xstar, xstarindex, fxstar = TTOpt(f,nodes,rmax,sweeps)

TensorTrainTools.TTOpt — Method

Finds the maximum element in an extended tensor train.

Signature

xstar, indxstar, fxstar = TTOpt(train,nodes)

TensorTrainTools.TTSIRT_GC — Method

sample = TTSIRTGC(train,N,domain,seed) sample = TTSIRTGC(train,N,domain)

Conditional distribution sampling from the square of a tensor train when the nodes are Gauss-Chebyshev.

TensorTrainTools.TTSIRT_GH — Method

sample = TTSIRTGH(train,N,seed) sample = TTSIRTGH(train,N)

Conditional distribution sampling from the square of a tensor train when the nodes are Gauss-Hermite.

TensorTrainTools.TTSIRT_GL — Method

sample = TTSIRTGL(train,N,domain,seed) sample = TTSIRTGL(train,N,domain)

Conditional distribution sampling from the square of a tensor train when the nodes are Gauss-Legendre.

TensorTrainTools.TTSIRT_PL — Method

sample = TTSIRTPL(train,N,domain,seed) sample = TTSIRTPL(train,N,domain)

Conditional distribution sampling from the square of a tensor train when the nodes are uniformly spaced.

TensorTrainTools.TTadd — Function

Add two discrete tensor trains. When the trains have different lengths, anchor determines whether the trains align from the first core (default) or the last core.

Signatures

t = TTadd(traina,trainb) t = TTadd(traina,trainb,anchor)

TensorTrainTools.TTals — Method

Tensor compression of the dense d-dimensional array, A.

Signatures

t = TTals(A,r,tol) t = TTals(A,r,tol,maxsweeps) t = TTals(A,r,tol,maxsweeps,seed)

TensorTrainTools.TTals — Method

Compute a discrete tensor train based on the function 'f' that populates the array.

Signature

t = TTals(f,nodes,r,tol) t = TTals(f,nodes,r,tol,maxsweeps) t = TTals(f,nodes,r,tol,maxsweeps,seed)

TensorTrainTools.TTcompute_marginal_GC — Method

Integrate a discrete tensor train over all dimensions except 'ind' using Gauss-Chebyshev quadrature.

Signature

margin = TTcomputemarginalGC(train,domain,ind)

TensorTrainTools.TTcompute_marginal_GH — Method

Integrate a discrete tensor train over all dimensions except 'ind' using Gauss-Hermite quadrature.

Signature

margin = TTcomputemarginalGH(train,ind)

TensorTrainTools.TTcompute_marginal_GL — Method

Integrate a discrete tensor train over all dimensions except 'ind' using Gauss-Legendre quadrature.

Signature

margin = TTcomputemarginalGL(train,domain,ind)

TensorTrainTools.TTcompute_marginal_PL — Method

Integrate a discrete tensor train over all dimensions except 'ind' using the trapazoidal method.

Signature

margin = TTcomputemarginalPL(train,domain,ind)

TensorTrainTools.TTconstant — Method

Create a tensor train of a constant 'value' with spacial dimensions 'n' and tensor ranks 'r' (an integer or vector of integers).

Signature

t = TTconstant(value,n,r)

TensorTrainTools.TTcreateCTT — Method

Create a Chebyshev tensor train from a discrete tensor train.

Signatures

trainCTT = TTcreateCTT(g,nodes,domain) trainCTT = TTcreateCTT(g,order,nodes,domain)

TensorTrainTools.TTcreateFTT — Method

Create a functional tensor train.

Signatures

trainFTT = TTcreateFTT(train,nodes,domain) # train is a discrete tensor train, nodes are Chebyshev roots trainFTT = TTcreateFTT(train,order,nodes,domain) # train is a discrete tensor train, nodes are Chebyshev roots trainFTT = TTcreateFTT(train,domain) # train is a ChebyshevTensorTrain or LegendreTensorTrain trainFTT = TTcreateFTT(train,order,domain) # train is a ChebyshevTensorTrain or LegendreTensorTrain trainFTT = TTcreateFTT(train,nodes) # train is a PiecewiseTensorTrain

TensorTrainTools.TTcreateLTT — Method

Create a Legendre tensor train from a discrete tensor train.

Signatures

trainLTT = TTcreateLTT(g,order,nodes,domain)

TensorTrainTools.TTcreatePTT — Method

Create a piecewise linear tensor train from a discrete tensor train.

Signature

trainPTT = TTcreatePTT(g)

TensorTrainTools.TTdecompress — Method

Recreates the dense array associated with a discrete tensor train.

Signature

A = TTdecompress(train)

TensorTrainTools.TTdecompress_threaded — Method

Recreates the dense array associated with a discrete tensor train.

Signature

A = TTdecompress_threaded(train)

TensorTrainTools.TTderivative — Method

Compute the derivative of a ChebyshevTensorTrain or LegendreTensorTrain with respect to one or two variables.

Signatures

deriv = TTderivative(train,x,pos,domain) deriv = TTderivative(train,x,pos1,pos2,domain)

TensorTrainTools.TTdescent — Method

Minimizes a continuous tensor train using gradient descent.

Signatures

newpoint, pointindex, fnewpoint, iters = TTdescent(train,nodes,point,domain) newpoint, pointindex, fnewpoint, iters = TTdescent(train,nodes,point,domain,α) newpoint, pointindex, fnewpoint, iters = TTdescent(train,nodes,point,domain,α,tol) newpoint, pointindex, fnewpoint, iters = TTdescent(train,nodes,point,domain,α,tol,maxiters)

TensorTrainTools.TTdescent — Method

Minimizes a discrete tensor train using gradient descent.

Signatures

newpoint, pointindex, fnewpoint, iters = TTdescent(train,nodes,point) newpoint, pointindex, fnewpoint, iters = TTdescent(train,nodes,point,α) newpoint, pointindex, fnewpoint, iters = TTdescent(train,nodes,point,α,tol) newpoint, pointindex, fnewpoint, iters = TTdescent(train,nodes,point,α,tol,maxiters) newpoint, pointindex, fnewpoint, iters = TTdescent(train,nodes,state,point) newpoint, pointindex, fnewpoint, iters = TTdescent(train,nodes,state,point,α) newpoint, pointindex, fnewpoint, iters = TTdescent(train,nodes,state,point,α,tol) newpoint, pointindex, fnewpoint, iters = TTdescent(train,nodes,state,point,α,tol,maxiters)

TensorTrainTools.TTdescent_step — Method

Perfoms one step of gradient descent on a discrete tensor train.

Signatures

newpoint, pointindex, fnewpoint, iters = TTdescentstep(train,nodes,point) newpoint, pointindex, fnewpoint, iters = TTdescentstep(train,nodes,point,α) newpoint, pointindex, fnewpoint, iters = TTdescentstap(train,nodes,state,point) newpoint, pointindex, fnewpoint, iters = TTdescentstap(train,nodes,state,point,α)

TensorTrainTools.TTevaluate — Method

Evaluate a discrete tensor train at an index of the dense array or a functional tensor train at a point.

Signatures

yhat = TTevaluate(train,index) yhat = TTevaluate(train,point)

TensorTrainTools.TTevaluate — Method

Evaluate a ChebyshevTensorTrain or LegendreTensorTrain at the point 'x'.

Signature

yhat = TTevaluate(train,x,domain)

TensorTrainTools.TTexp — Method

Approximate the exponential of a tensor train in tensor train format.

Signature

t = TTexp(train,N,tol)

TensorTrainTools.TTextremize — Method

Extremize over the cores of a tensor train. The output could be a maxima or a minima. If 'state' is provided, the initial cores in the tensor train are associated with the state variables. 'state' can be an index for a grid location or a point in the state space.

Signatures

soln = TTextremize(train,K) soln = TTextremize(train,K,state) # state is a vector of integers soln = TTextremize(train,K,nodes,state) # state is a vector of floating point numbers

TensorTrainTools.TTgradient — Method

Compute the gradient of a tensor train.

Signature

grad = TTgradient(train,x,domain) # ChebyshevTensorTrain or LegendreTensorTrain grad = TTgradient(train,nodes,x) # DiscreteTensorTrain

TensorTrainTools.TThessian — Method

Compute the Hessian of a tensor train.

Signature

hess = TThessian(train,x,domain) # ChebyshevTensorTrain or LegendreTensorTrain hess = TThessian(train,nodes,x) # DiscreteTensorTrain

TensorTrainTools.TTinner_prod — Method

Compute the inner product of two tensor trains.

Signature

innerprod = TTinnerprod(traina,trainb)

TensorTrainTools.TTinner_prod2 — Method

Compute the inner product of two tensor trains where traina has fewer dimensions than trainb.

Signature

trainc = TTinner_prod2(traina,trainb)

TensorTrainTools.TTintegrate_GC — Method

Integrate a discrete tensor train using Gauss-Chebyshev quadrature.

Signatures

area = TTintegrateGC(train,domain) # Integrates over all dimensions area = TTintegrateGC(train,domain,μ) # Integrates over dimensions 1 to μ

TensorTrainTools.TTintegrate_GC — Method

Integrate a functional tensor train over all dimensions using Gauss-Chebyshev quadrature.

Signature

area = TTintegrate_GC(train,nodes,domain)

TensorTrainTools.TTintegrate_GH — Method

Integrate a discrete tensor train using Gauss-Hermite quadrature.

Signatures

area = TTintegrateGH(train) # Integrates over all dimensions area = TTintegrateGH(train,μ) # Integrates over dimensions 1 to μ

TensorTrainTools.TTintegrate_GH — Method

Integrate a functional tensor train over all dimensions using Gauss-Hermite quadrature.

Signature

area = TTintegrate_GH(train,nodes)

TensorTrainTools.TTintegrate_GL — Method

Integrate a discrete tensor train using Gauss-Legendre quadrature.

Signatures

area = TTintegrateGL(train,domain) # Integrates over all dimensions area = TTintegrateGL(train,domain,μ) # Integrates over dimensions 1 to μ

TensorTrainTools.TTintegrate_LC — Method

Integrate a functional tensor train over all dimensions using Gauss-Legendre quadrature.

Signature

area = TTintegrate_GL(train,nodes,domain)

TensorTrainTools.TTintegrate_PL — Method

Integrate a discrete tensor train using the trapazoidal method.

Signatures

area = TTintegratePL(train,domain) # Integrates over all dimensions area = TTintegratePL(train,domain,μ) # Integrates over dimensions 1 to μ

TensorTrainTools.TTinterp — Method

Create an interpolating function from a functional tensor train.

Signature

f = TTinterp(train)

TensorTrainTools.TTmult — Method

Multiply a discrete tensor train and a scalar.

Signatures

t = TTmult(train,s) t = TTmult(s,train)

TensorTrainTools.TTnewton — Method

Minimizes a continuous tensor train using Newton's method.

Signature

newpoint, pointindex, fnewpoint, iters = TTnewton(train,nodes,point,tol,maxiters)

TensorTrainTools.TTnewton — Method

Minimizes a discrete tensor train using Newton's method.

Signatures

newpoint, pointindex, fnewpoint, iters = TTnewton(train,nodes,point,tol,maxiters) newpoint, pointindex, fnewpoint, iters = TTnewton(train,nodes,state,point,tol,maxiters)

TensorTrainTools.TTnewton_step — Method

Perfoms one step of Newton's method on a discrete tensor train.

Signatures

newpoint, pointindex, fnewpoint, iters = TTnewtonstep(train,nodes,point,tol,maxiters) newpoint, pointindex, fnewpoint, iters = TTnewtonstap(train,nodes,state,point,tol,maxiters)

TensorTrainTools.TTnorm — Method

Compute the norm of a discrete tensor train.

Signature

TTnorm(train)

TensorTrainTools.TTones — Method

Create a tensor train of a ones of eltype(T) with spacial dimensions 'n' and tensor ranks 'r' (an integer or vector of integers).

Signatures

t = TTones(n,r) t = TTones(n,r,T)

TensorTrainTools.TToptimize — Method

Optimize over the cores of a tensor train, returning the maximum and the minimum. If 'state' is provided, the initial cores in the tensor train are associated with the state variables. 'state' can be an index for a grid location or a point in the state space.

Signatures

(imin,imax),(ymin,ymax) = TToptimize(train,K,tol) (imin,imax),(ymin,ymax) = TToptimize(train,K,state,tol) # state is a vector of integers (imin,imax),(ymin,ymax) = TToptimize(train,K,nodes,state,tol) # state is a vector of floating point numbers

TensorTrainTools.TTorthleft — Method

Left orthogonalise a discrete tensor train.

Signature

t = TTorthleft(train)

TensorTrainTools.TTorthleftright — Method

Left orthogonalise a train up to core 'μ' and right orthogonalise the remaining cores.

Signature

t = TTorthleftright(train,μ)

TensorTrainTools.TTorthright — Method

Right orthogonalise a discrete tensor train.

Signature

t = TTorthright(train)

TensorTrainTools.TTpower — Method

Construct the N'th power (element-by-element) of a tensor train.

Signature

t = TTpower(train,N)

TensorTrainTools.TTrand — Method

Create a tensor train with uniformly random entries of eltype(T) that has spacial dimensions 'n' and tensor ranks 'r' (an integer or vector of integers).

Signatures

t = TTrand(n,r) t = TTrand(n,r,T)

TensorTrainTools.TTrandn — Method

Create a tensor train with normally distributed random entries of eltype(T) that has spacial dimensions 'n' and tensor ranks 'r' (an integer or vector of integers).

Signatures

t = TTrandn(n,r) t = TTrandn(n,r,T)

TensorTrainTools.TTreverse_indices — Method

Reverse the cores in a discrete tensor train (reversing the order of the variables).

Signature

t = Ttreverse(train)

TensorTrainTools.TTrounding — Method

Perform rounding on a discrete tensor train to find a more compact discrete tensor train representation.

Signature

t = TTrounding(train,tol)

TensorTrainTools.TTsize — Method

Compute the (implied) size of a decompressed tensor train.

Signature

n = TTsize(train)

TensorTrainTools.TTsquared — Method

Construct the square (element-by-element) of a tensor train.

Signature

t = TTsquared(train)

TensorTrainTools.TTsubtract — Function

Subtract two discrete tensor trains. When the trains have different lengths, anchor determines whether the trains align from the first core (default) or the last core.

Signatures

t = TTsubstract(traina,trainb) t = TTsubstract(traina,trainb,anchor)

TensorTrainTools.TTsvd — Method

Tensor compression of the dense d-dimensional array, A.

Signatures

t = TTsvd(A,tol) t = TTsvd(A,r)

TensorTrainTools.TTsvd — Method

Compute a discrete tensor train based on the function 'f' that populates the array.

Signature

t = TTsvd(f,nodes,tol)

TensorTrainTools.TTsvd_threaded — Method

Compute a discrete tensor train based on the function 'f' that populates the array.

Signature

t = TTsvd_threaded(f,nodes,tol)

TensorTrainTools.TTzeros — Method

Create a tensor train of zeros of eltype(T) with spacial dimensions 'n' and tensor ranks 'r'.

Signatures

t = TTzeros(n,r) t = TTzeros(n,r,T)

TensorTrainTools.bisection — Method

Use bisection to find the fix-point of the continuous function 'f'.

Signature

xstar, f_xstar, iters = bisection(f,x,tol,maxiters)

TensorTrainTools.bracket_nodes — Method

Finds the indexes for the elements of 'x' that immediately bracket 'point'.

Signature

(l,u) = bracket_nodes(x,point)

TensorTrainTools.cheb_evaluate — Method

Evaluate a Chebyshev polynomial at point, 'point'.

Signature

yhat = cheb_evaluate(w,point,order,domain)

TensorTrainTools.cheb_nodes — Method

Create a vector of 'n' Chebyshev nodes on 'domain' with element type 'T'.

Signatures

nodes = chebnodes(n) nodes = chebnodes(n,domain) nodes = cheb_nodes(n,domain,T)

TensorTrainTools.cheb_polynomial — Method

Generate Chebyshev polynomials of 'order' at 'point'. 'point' can be a scalar or a vector of scalars.

Signature

P = chebyshev_polynomial(order,point)

TensorTrainTools.cheb_polynomial_deriv — Method

Construct the derivative of a Chebyshev polynomial at 'point'. 'point' can be a scalar or a vector of scalars.

Signature

deriv = chebpolynomialderiv(order,point)

TensorTrainTools.cheb_weights — Method

Compute the weights of a univariate Chebyshev polynomial with order, 'order', on domain, 'domain'.

Signature

w = cheb_weights(y,nodes,order,domain)

TensorTrainTools.chebyshev_interp — Method

Create a function that evaluates a Chebyshev polynomial at point, 'point'.

f = chebyshev_interp(y,nodes,order,domain)

TensorTrainTools.ind2sub — Method

Find the array sub-index from a linear index; returns a tuple or tuple of tuples if 'i' is a vector.

Signature

s = ind2sub(i,dims)

TensorTrainTools.invert_cdf — Method

Take a draw from a density by inverting the CDF.

Signatures

c = invertcdf(f,u) # Unbounded domain c = invertcdf(f,u,lb,ub) # Bounded domain

TensorTrainTools.legendre_evaluate — Method

Evaluate a Legendre polynomial at point, 'point'.

Signature

yhat = legendre_evaluate(w,point,order,domain)

TensorTrainTools.legendre_interp — Method

Create a function that evaluates a Legendre polynomial at point, 'point'.

Signature

f = legendre_interp(y,nodes,order,domain)

TensorTrainTools.legendre_nodes — Method

Create a vector of 'n' Legendre nodes on 'domain' with element type 'T'.

Signatures

nodes= legendrenodes(n) nodes= legendrenodes(n,domain) nodes= legendre_nodes(n,domain,T)

TensorTrainTools.legendre_polynomial — Method

Generate Legendre polynomials of 'order' at 'point'. 'point' can be a scalar or a vector of scalars.

Signatures

P = legendre_polynomial(order,point)

TensorTrainTools.legendre_polynomial_deriv — Method

Construct the derivative of a Legendre polynomial at 'point'. 'point' can be a scalar or a vector of scalars.

Signature

deriv = legendrepolynomialderiv(order,point)

TensorTrainTools.legendre_polynomial_sec_deriv — Method

Construct the second derivative of a Legendre polynomial at 'point'. ' point can be a scalar or a vector of scalars.

Signature

deriv = legendrepolynomialderiv(order,point)

TensorTrainTools.legendre_weights — Method

Compute the weights of a univariate Legendre polynomial with order, 'order', on domain, 'domain'.

Signature

w = legendre_weights(y,nodes,order,domain)

TensorTrainTools.normalize_node — Method

Normalize 'node' so that it resides within the interval [-1,1]. 'node' can be a scalar or a vector.

Signature

newnode = normalizenode(node,domain)

TensorTrainTools.piecewise_linear_evaluate — Method

Create a function that evaluates a piecewise linear approximation at point, 'point'.

Signature

f = piecewiselinearevaluate(y,nodes)

TensorTrainTools.piecewise_linear_nodes — Method

Compute 'n' points uniformly placed on domain, 'domain', with eltype(T)

Signatures

nodes = piecewiselinearnodes(n) nodes = piecewiselinearnodes(n,domain) nodes = piecewiselinearnodes(n,domain,T)

TensorTrainTools.rq — Method

Perform the RQ matrix decomposition.

Signature

R,Q = rq(A)

TensorTrainTools.stack — Method

Horizontally concatenate two matrices.

Signature

c = stack(a,b)

TensorTrainTools.times_cores — Method

Multiply two tensor train cores (3D arrays) to produce a 4D array.

Signature

t = times_cores(A,B)

TensorTrainTools.times_dim_1 — Method

Pre-multiply a 3d array, B, by a matrix, A, along the first dimension of the 3D array to produce a 3D array.

Signature

t = timesdim1(A,B)

TensorTrainTools.times_dim_1 — Method

Pre-multiply a 4D array, B, by a matrix, A, along the first dimension of the 4D array to produce a 4D array.

Signature

t = timesdim1(A,B)

TensorTrainTools.times_dim_3 — Method

Pre-multiply a matrix, B, by a 3D array, A, along the third dimension of the 3D array to produce a 3D array.

Signature

t = timesdim3(A,B)

TensorTrainTools.times_dim_4 — Method

Pre-multiply a matrix, B, by a 4D array, A, along the fourth dimension of the 4D array to produce a 4D array.

Signature

t = timesdim4(A,B)

TensorTrainTools.topK — Method

Find the 'K' rows of a matrix with the highest norm.

Signature

p = topK(M,K)

TensorTrainTools.trapazoidal — Method

Use the trapazoidal method to integrate the cores of a discrete tensor train.

Signature

integral = trapazoidal(y,domain)

TensorTrainTools.tsvd — Method

Compute the truncated SVD decomposition of a matrix with singular value threshold 'δ'.

Signatures

t = tsvd(A,δ) t = tsvd(A,δ,r)