TensorTrainApprox

Documentation for TensorTrainApprox.

TensorTrainApprox.DMRGcross

TensorTrainApprox.DMRGcross

TensorTrainApprox.DMRGcross_generic

TensorTrainApprox.DMRGcross_generic

TensorTrainApprox.DMRGcross_threaded

TensorTrainApprox.TTderivative

TensorTrainApprox.TTderivative

TensorTrainApprox.TTevaluate

TensorTrainApprox.TTevaluate

TensorTrainApprox.TTevaluate

TensorTrainApprox.TTevaluate

TensorTrainApprox.TTgradient

TensorTrainApprox.TTgradient

TensorTrainApprox.TTintegrate_GC

TensorTrainApprox.TTintegrate_GC

TensorTrainApprox.TTintegrate_GH

TensorTrainApprox.TTintegrate_GH

TensorTrainApprox.TTintegrate_GL

TensorTrainApprox.TTinterp

TensorTrainApprox.TTsvd

TensorTrainApprox.chebyshev_evaluate

TensorTrainApprox.chebyshev_interp

TensorTrainApprox.chebyshev_nodes

TensorTrainApprox.chebyshev_polynomial

TensorTrainApprox.chebyshev_polynomial

TensorTrainApprox.chebyshev_polynomial_deriv

TensorTrainApprox.chebyshev_polynomial_deriv

TensorTrainApprox.chebyshev_weights

TensorTrainApprox.constantTT

TensorTrainApprox.constantTT

TensorTrainApprox.createCTT

TensorTrainApprox.createCTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createFTT

TensorTrainApprox.createLTT

TensorTrainApprox.createLTT

TensorTrainApprox.createPTT

TensorTrainApprox.decompress

TensorTrainApprox.decompress_threaded

TensorTrainApprox.frobenius

TensorTrainApprox.frobenius

TensorTrainApprox.ind2sub

TensorTrainApprox.ind2sub

TensorTrainApprox.legendre_evaluate

TensorTrainApprox.legendre_interp

TensorTrainApprox.legendre_nodes

TensorTrainApprox.legendre_polynomial

TensorTrainApprox.legendre_polynomial

TensorTrainApprox.legendre_polynomial_deriv

TensorTrainApprox.legendre_polynomial_deriv

TensorTrainApprox.legendre_weights

TensorTrainApprox.normalize_node

TensorTrainApprox.normalize_node

TensorTrainApprox.piecewise_linear_nodes

TensorTrainApprox.randomTT

TensorTrainApprox.randomTT

TensorTrainApprox.times_cores

TensorTrainApprox.tsvd

TensorTrainApprox.DMRGcross — Method

Compute a tensor train approximation of a dense d-dimensional array. DMRG tensor compression based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross — Method

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_generic — Method

Compute a tensor train approximation of a dense d-dimensional array. DMRG tensor compression based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_generic — Method

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.DMRGcross_threaded — Method

Compute a tensor train approximation of a dense d-dimensional array based on the function that populates the array. DMRG function approximation based on Dolgov and Savostyanov (2020).

TensorTrainApprox.TTderivative — Method

Compute the derivative of a Chebyshev tensor train with respect to a variable.

TensorTrainApprox.TTderivative — Method

Compute the derivative of a Legendre tensor train with respect to a variable.

TensorTrainApprox.TTevaluate — Method

Evaluate a tensor train at a vector of index points of the dense array.

TensorTrainApprox.TTevaluate — Method

Evaluate a tensor train at an index point of the dense array.

TensorTrainApprox.TTevaluate — Method

Evaluate a functional tensor train at a vector of 'points'.

TensorTrainApprox.TTevaluate — Method

Evaluate a functional tensor train at 'point'.

TensorTrainApprox.TTgradient — Method

Compute the gradient vector of a Chebyshev tensor train.

TensorTrainApprox.TTgradient — Method

Compute the gradient vector of a Legendre tensor train.

TensorTrainApprox.TTintegrate_GC — Method

Integrate a tensor train using Gauss-Chebyshev quadrature.

TensorTrainApprox.TTintegrate_GC — Method

Integrate a functional tensor train using Gauss-Legendre quadrature.

TensorTrainApprox.TTintegrate_GH — Method

Integrate a tensor train using Gauss-Hermite quadrature.

TensorTrainApprox.TTintegrate_GH — Method

Integrate a functional tensor train using Gauss-Hermite quadrature.

TensorTrainApprox.TTintegrate_GL — Method

Integrate a tensor train using Gauss-Legendre quadrature.

TensorTrainApprox.TTinterp — Method

Create an interpolating function from a functional tensor train.

TensorTrainApprox.TTsvd — Method

Compute a tensor train approximation of a dense d-dimensional array. Tensor compression based on Oseledets and Tyrtyshnikov (2010).

TensorTrainApprox.chebyshev_evaluate — Method

Evaluate a Chebyshev polynomial at 'point'.

TensorTrainApprox.chebyshev_interp — Method

Create a function that evaluates a Chebyshev polynomial at 'point'.

TensorTrainApprox.chebyshev_nodes — Method

Create a vector of 'n' Chebyshev nodes on 'domain' with element type 'T'.

TensorTrainApprox.chebyshev_polynomial — Method

Generate Chebyshev polynomials of 'order' at 'node'.

TensorTrainApprox.chebyshev_polynomial — Method

Generate Chebyshev polynomials of 'order' at each element of 'nodes'.

TensorTrainApprox.chebyshev_polynomial_deriv — Method

Construct the derivative of a Chebyshev polynomial at 'x'.

TensorTrainApprox.chebyshev_polynomial_deriv — Method

Construct the derivative of a Chebyshev polynomial at each element of 'x'.

TensorTrainApprox.chebyshev_weights — Method

Compute the weights for a Chebyshev polynomial.

TensorTrainApprox.constantTT — Method

Create a tensor train of a constant 'value' with spacial dimensions 'n' and tensor ranks 'r'.

TensorTrainApprox.constantTT — Method

Create a tensor train of a constant 'value' with spacial dimensions 'n' and tensor ranks 'r'.

TensorTrainApprox.createCTT — Method

Create a Chebyshev tensor train from a tensor train.

TensorTrainApprox.createCTT — Method

Create a Chebyshev tensor train from a tensor train.

TensorTrainApprox.createFTT — Method

Create a functional tensor train from a Chebyshev tensor train.

TensorTrainApprox.createFTT — Method

Create a functional tensor train from a Legendre tensor train.

TensorTrainApprox.createFTT — Method

Create a functional tensor train from a tensor train, assuming the nodes are Chebyshev roots.

TensorTrainApprox.createFTT — Method

Create a functional tensor train from a Chebyshev tensor train.

TensorTrainApprox.createFTT — Method

Create a functional tensor train from a Legendre tensor train.

TensorTrainApprox.createFTT — Method

Create a functional tensor train from a tensor train, assuming the nodes are Chebyshev roots.

TensorTrainApprox.createFTT — Method

Create a functional tensor train from a piecewise tensor train.

TensorTrainApprox.createLTT — Method

Create a Legendre tensor train from a tensor train.

TensorTrainApprox.createLTT — Method

Create a Legendre tensor train from a tensor train.

TensorTrainApprox.createPTT — Method

Create a piecewise linear tensor train from a tensor train.

TensorTrainApprox.decompress — Method

Recreates the dense array from a tensor train.

TensorTrainApprox.decompress_threaded — Method

Recreates the dense array from a tensor train.

TensorTrainApprox.frobenius — Method

Compute the Frobenius norm of a d-dimensional array.

TensorTrainApprox.ind2sub — Method

Find the array sub-indices from a linear index.

TensorTrainApprox.ind2sub — Method

Find the array sub-indices from a linear index.

TensorTrainApprox.legendre_evaluate — Method

Evaluate a Legendre polynomial at 'point'.

TensorTrainApprox.legendre_interp — Method

Create a function that evaluates a Legendre polynomial at 'point'.

TensorTrainApprox.legendre_nodes — Method

Create a vector of 'n' Legendre nodes on 'domain' with element type 'T'.

TensorTrainApprox.legendre_polynomial — Method

Generate Legendre polynomials of 'order' at 'node'.

TensorTrainApprox.legendre_polynomial — Method

Generate Legendre polynomials of 'order' at each element of 'nodes'.

TensorTrainApprox.legendre_polynomial_deriv — Method

Construct the derivative of a Legendre polynomial at 'x'.

TensorTrainApprox.legendre_polynomial_deriv — Method

Construct the derivative of a Legendre polynomial at each element of 'x'.

TensorTrainApprox.legendre_weights — Method

Compute the weights for a Legendre polynomial.

TensorTrainApprox.normalize_node — Method

Normalize 'node' so that it resides within [1,-1].

TensorTrainApprox.normalize_node — Method

Normalize 'node' so that all elements reside within [1,-1].

TensorTrainApprox.piecewise_linear_nodes — Method

Create a vector of 'n' uniformly spaced nodes on 'domain' with element type 'T'.

TensorTrainApprox.randomTT — Method

Create a random tensor train with spacial dimensions 'n' and tensor ranks 'r'.

TensorTrainApprox.randomTT — Method

Create a random tensor train with spacial dimensions 'n' and tensor ranks 'r'.

TensorTrainApprox.times_cores — Method

Multiply two cores of a tensor train.

TensorTrainApprox.tsvd — Method

Compute the truncated SVD decomposition of a matrix.

TensorTrainApprox.frobenius — Method

Compute the Frobenius norm of a d-dimensional array.